Implemented flask`s blueprints
+Added blueprint for `registration` route in separate file `blueprints.py` in
`func` dir.
-Deleted `registration` route from `app.py`
-Deleted configuration overhaul from `test_app.py` in `tests` dir.
+Fixed some blank lines.
