+++ api/func/__init__.py
@@ -3,14 +3,22 @@ from pathlib import Path
 from flask import Flask
 from flask_migrate import Migrate
 
-from config import BaseConfig
-from func.database import db, User
+from func.blueprints import registration_blueprint
+from config import BaseConfig, TestConfig
+from func.database import db
 
 
-def create_app():
-    app = Flask(__name__, instance_relative_config=True)
+def create_app(test_config=False):
+    app = Flask(__name__)
 
-    app.config.from_object(BaseConfig())
+    if test_config is True:
+        app.config.from_object(TestConfig())
+        print('test')
+    else:
+        app.config.from_object(BaseConfig())
+        print('base')
+
+    app.register_blueprint(registration_blueprint)
 
     db.init_app(app)
     migrate = Migrate(app, db)
