+++ api/run_tests.py
@@ -1,5 +1,6 @@
 if __name__ == '__main__':
     import unittest
+
     tests = unittest.TestLoader().discover('tests')
     result = unittest.TextTestRunner(verbosity=2).run(tests)
